chore: rename `utils/address` to `utils/districts`
